rmcp/requirements.txt
# RMCP Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
pyyaml==6.0.1
# sqlite3 is built into Python
numpy==1.24.3
scikit-learn==1.3.2
openai==1.3.7
anthropic==0.7.8
python-multipart==0.0.6

# Phase 2 Dependencies
sentence-transformers==2.2.2  # For local embeddings
torch==2.1.0  # Required by sentence-transformers
transformers==4.35.0  # Required by sentence-transformers

# Phase 3 Dependencies - Security & Production
PyJWT==2.8.0  # JWT token handling
cryptography==41.0.7  # Cryptographic functions
python-jose[cryptography]==3.3.0  # JWT and JWE support

# Phase 3.4 Dependencies - Observability
prometheus-fastapi-instrumentator==6.1.0  # Prometheus metrics for FastAPI
prometheus-client==0.19.0  # Prometheus client library

# Phase 3.6 Dependencies - Structured Logging
structlog==23.2.0  # Structured logging
python-json-logger==2.0.7  # JSON log formatter

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
